Fixed pain sounds for female peds (they used to make male pain noises)

+++ trunk/MTA10/game_sa/CPedModelInfoSA.h
@@ -43,7 +43,7 @@ class CPedModelInfoSAInterface : public CBaseModelInfoSAInterface
 public:
     BYTE                pad [ 4 ];              // 32   
     AssocGroupID        motionAnimGroup;        // 36   Motion anim group (AssocGroupID, long)
-    long                pedType;                // 40   Default ped type (long)
+    DWORD               pedType;                // 40   Default ped type (long)
     ePedStats *         pedStats;               // 44   Default ped stats (ePedStats)
     BYTE                bCanDriveCars;          // 48   Can drive cars (byte)
     BYTE                pad2 [ 1 ];             // 49
